+++ generators/dashboard/templates/packages/components/src/Failure.js
@@ -1,7 +1,5 @@
 import React from 'react';
 
-const Failure = ({ error, refetch }) => {
+export default function Failure({ error, refetch }) {
   return <div className="failure">Loading failure</div>;
-};
-
-export default Failure;
+}