+++ generators/dashboard/templates/packages/components/src/PageHeader.js
@@ -1,12 +1,10 @@
 import React from 'react';
 
-const PageHeader = ({ title, extra }) => {
+export default function PageHeader({ title, extra }) {
   return (
     <div className="page-header">
       <div className="page-header-title">{title}</div>
       {extra}
     </div>
   );
-};
-
-export default PageHeader;
+}