+++ generators/dashboard/templates/packages/dashboard/src/layouts/Layout.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import { Sidebar } from 'components';
 
-const Layout = ({ children }) => {
+export default function Layout({ children }) {
   return (
     <div className="app">
       <Sidebar />
       <div className="content">{children}</div>
     </div>
   );
-};
-
-export default Layout;
+}