+++ generators/dashboard/templates/packages/dashboard/src/pages/home.js
@@ -1,7 +1,5 @@
 import React from 'react';
 
-const Home = () => {
+export default function Home() {
   return <div>Home</div>;
-};
-
-export default Home;
+}